refactor(Docs): render processors through the shared sections
The sections plumbing was copied in three generators; it now lives on
DocGenerator, so augmenters and spec attributes regenerate unchanged.
processors.md moves from inline HTML to the same markdown the other
pages use, and ConfigSettingsSection keeps multi-line descriptions
inside their list item.
